Exampleageddon Greatly exapanded normal sound examples Introduced historical sound examples Brought in many historical .a and .dll files for examples cleaned up compile warnings on win32/64 msys, cygwin mingw versions of xbox, android, js build scripts resolved rsync excludes Successfully compiled tests on mingw/cygwin Modified makefiles to fit with the others

+++ src/rtlib/win32/sys_execex.c
@@ -92,7 +92,7 @@ FBCALL int fb_ExecEx( FBSTRING *program, FBSTRING *args, int do_fork )
                 }
                 CloseHandle( ProcessInfo.hProcess );
             } else {
-                res = (int) ProcessInfo.hProcess;
+                res = (int)(intptr_t)ProcessInfo.hProcess;
             }
         }
 #endif
